app/frontend/requirements.txt
flask==3.0.0
gunicorn==21.2.0
prometheus-flask-exporter==0.23.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
